Move out tutorial and finish button.
